+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 * **`1.0.0`**
     * **Use GATs for `ArrayLength`** !
     * Bump MSRV to 1.65.0
-    * Use Rust 2021 edition [#118](https://github.com/fizyk20/generic-array/pull/118)
+    * Use Rust 2021 edition [#118](https://github.com/fizyk20/generic-array/pull/118) + non-PR upgrade later with GATs.
     * Allow `arr!` macro in `const` [#129](https://github.com/fizyk20/generic-array/pull/129)
     * Add `arr!` repeat-expressions [#130](https://github.com/fizyk20/generic-array/pull/130)
     * Implement `const-default` trait support [#131](https://github.com/fizyk20/generic-array/pull/131)
     * Make `as_slice()` const
+    * Make `ArrayLength: 'static`
 
 * **`0.14.6`**
     * Add an optional `Zeroize` impl for `GenericArray` ([#126](https://github.com/fizyk20/generic-array/pull/126) and [#112](https://github.com/fizyk20/generic-array/pull/112))